refactor(auth): pattern cleanups from identity-linking review
- #610: insert_user_identity now wraps its terminal .await in db_error!,
  matching the logging convention used across db/queries.rs.
- #611: remove count_user_identities — the atomic-unlink rewrite inlined its
  own COUNT, leaving it test-only. The two tests now assert via
  list_user_identities(...).len().
- #612: de-duplicate the compact relative-time formatter. SessionsSection and
  LinkedAccountsSection each had an identical local copy; extracted as
  utils::formatRelativeTimeShort (distinct from the existing verbose
  formatRelativeTime, kept so the compact "5m ago" display is unchanged).

Verified: clippy -D warnings clean, nightly fmt clean, 52 server identity/oidc
tests pass, 606 client tests pass, bun run build clean.

Closes #610, closes #611, closes #612.
